Update to load local overrides only (if found)
This lets local files override or modify the main sourced file
as they need to - they're not stuck always coming in after
the base file is loaded, if they need to run a different way.

To update, move all _local files to local/${filename%_local}
and add a `load-global-config "$BASH_SOURCE" to the top
(this will get the same load order you had *before* this change.